@@ -58,56 +66,99 @@
     "passport-jwt": "^4.0.0",
     "pg": "^8.8.0",
     "plaid": "^11.0.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-google-autocomplete": "^2.7.0",
+    "react-hook-form": "^7.38.0",
+    "react-i18next": "^11.18.6",
+    "react-plaid-link": "^3.3.2",
+    "react-qr-code": "^2.0.8",
+    "react-query": "^3.39.2",
+    "react-router-dom": "^6.4.2",
+    "react-table": "^7.8.0",
+    "react-tooltip": "^4.4.2",
     "redlock": "^5.0.0-beta.2",
     "reflect-metadata": "^0.1.13",
+    "regenerator-runtime": "0.13.10",
     "rxjs": "^7.5.7",
-    "supertest": "^6.2.4",
+    "styled-components": "^5.3.6",
+    "supertest": "^6.3.0",
     "tslib": "^2.4.0",
     "typeorm": "^0.3.10",
+    "use-async-effect": "^2.2.7",
     "uuid": "^9.0.0",
     "validator": "^13.7.0",
-    "winston": "^3.8.2"
+    "winston": "^3.8.2",
+    "yup": "^0.32.11"
   },
   "devDependencies": {
+    "@babel/core": "7.19.6",
+    "@babel/preset-typescript": "7.18.6",
     "@nestjs/schematics": "^9.0.3",
-    "@nestjs/testing": "^9.1.2",
-    "@nrwl/cli": "14.8.2",
-    "@nrwl/devkit": "14.8.2",
-    "@nrwl/eslint-plugin-nx": "14.8.2",
-    "@nrwl/jest": "14.8.2",
-    "@nrwl/js": "14.8.2",
-    "@nrwl/linter": "14.8.2",
-    "@nrwl/nest": "14.8.2",
-    "@nrwl/node": "^14.8.2",
-    "@nrwl/workspace": "14.8.2",
-    "@swc-node/register": "^1.5.2",
-    "@swc/cli": "~0.1.57",
-    "@swc/core": "^1.3.4",
-    "@types/auth0": "^2.35.5",
-    "@types/jest": "29.1.1",
+    "@nestjs/testing": "^9.1.4",
+    "@nrwl/cli": "15.0.0",
+    "@nrwl/devkit": "15.0.0",
+    "@nrwl/eslint-plugin-nx": "15.0.0",
+    "@nrwl/jest": "15.0.0",
+    "@nrwl/js": "15.0.0",
+    "@nrwl/linter": "15.0.0",
+    "@nrwl/nest": "15.0.0",
+    "@nrwl/node": "15.0.0",
+    "@nrwl/nx-cloud": "^15.0.0",
+    "@nrwl/react": "15.0.0",
+    "@nrwl/storybook": "15.0.0",
+    "@nrwl/web": "15.0.0",
+    "@nrwl/workspace": "15.0.0",
+    "@react-theming/storybook-addon": "^1.1.7",
+    "@storybook/addon-essentials": "6.5.12",
+    "@storybook/builder-webpack5": "6.5.12",
+    "@storybook/manager-webpack5": "6.5.12",
+    "@storybook/react": "6.5.12",
+    "@svgr/webpack": "6.5.0",
+    "@types/auth0": "^2.35.6",
+    "@types/jest": "29.2.0",
     "@types/jest-when": "^3.5.2",
     "@types/lodash": "^4.14.186",
-    "@types/luxon": "^3.0.1",
+    "@types/luxon": "^3.0.2",
     "@types/node": "16.11.7",
     "@types/passport-jwt": "^3.0.7",
+    "@types/react": "^18.0.21",
+    "@types/react-dom": "^18.0.6",
+    "@types/react-router-dom": "5.3.3",
+    "@types/react-table": "^7.7.12",
+    "@types/styled-components": "^5.1.26",
     "@types/supertest": "^2.0.12",
     "@types/uuid": "^8.3.4",
-    "@typescript-eslint/eslint-plugin": "~5.38.1",
-    "@typescript-eslint/parser": "~5.38.1",
-    "chalk": "5.0.1",
+    "@typescript-eslint/eslint-plugin": "~5.40.1",
+    "@typescript-eslint/parser": "~5.40.1",
+    "@swc-node/register": "^1.5.4",
+    "@swc/core": "^1.3.10",
+    "babel-jest": "29.2.1",
+    "babel-loader": "8.2.5",
+    "babel-plugin-styled-components": "^2.0.7",
     "clear": "^0.1.0",
     "commander": "^9.4.1",
-    "eslint": "~8.24.0",
+    "cypress": "^10.10.0",
+    "eslint": "~8.25.0",
     "eslint-config-prettier": "8.5.0",
-    "jest": "29.1.2",
-    "jest-environment-jsdom": "^29.1.2",
+    "eslint-plugin-cypress": "^2.12.1",
+    "eslint-plugin-import": "2.26.0",
+    "eslint-plugin-jsx-a11y": "6.6.1",
+    "eslint-plugin-react": "7.31.10",
+    "eslint-plugin-react-hooks": "4.6.0",
+    "husky": "^8.0.1",
+    "jest": "29.2.1",
     "jest-when": "^3.5.1",
     "jsonc-eslint-parser": "^2.1.0",
-    "nx": "14.8.2",
+    "nx": "15.0.0",
     "prettier": "^2.7.1",
+    "storybook-addon-react-router-v6": "^0.2.1",
     "ts-jest": "29.0.3",
     "ts-node": "~10.9.1",
     "tslib": "^2.4.0",
-    "typescript": "~4.8.4"
+    "typescript": "~4.8.4",
+    "url-loader": "^4.1.1",
+    "webpack": "^5.74.0",
+    "webpack-merge": "^5.8.0"
   }
 }
