@@ -1,7 +1,6 @@
 /* eslint-disable */
 
 import { Injectable, Logger } from '@nestjs/common';
-import { GetKycPayload, GetKycResponse } from '@archie/api/user-api/kyc';
 import { RizeApiService } from './api/rize_api.service';
 import {
   ComplianceDocumentAcknowledgementRequest,
@@ -30,10 +29,6 @@ import {
   GET_USER_KYC_RPC,
 } from '@archie/api/user-api/constants';
 import {
-  GetEmailAddressPayload,
-  GetEmailAddressResponse,
-} from '@archie/api/user-api/user';
-import {
   CardActivatedPayload,
   TransactionUpdatedPayload,
 } from '@archie/api/credit-api/data-transfer-objects';
@@ -48,7 +43,13 @@ import {
 import { LessThanOrEqual, Repository, UpdateResult } from 'typeorm';
 import { InjectRepository } from '@nestjs/typeorm';
 import { LastDebitCardUpdateMeta } from './last_debit_card_update_meta.entity';
-import { Lock } from '@archie-microservices/api/utils/redis';
+import { Lock } from '@archie/api/utils/redis';
+import {
+  GetEmailAddressPayload,
+  GetEmailAddressResponse,
+  GetKycPayload,
+  GetKycResponse,
+} from '@archie/api/user-api/data-transfer-objects';
 
 @Injectable()
 export class RizeService {
@@ -142,7 +143,7 @@ export class RizeService {
       const transaction: RizeTransaction =
         await this.rizeApiService.getTransaction(transactionId);
 
-      this.queueService.publish<TransactionUpdatedPayload>(
+      this.queueService.publishEvent(
         TRANSACTION_UPDATED_TOPIC,
         {
           ...transaction,
@@ -289,7 +290,7 @@ export class RizeService {
       customerId,
       complianceWorkflow.product_uid,
     );
-    this.queueService.publish<CardActivatedPayload>(CARD_ACTIVATED_TOPIC, {
+    this.queueService.publishEvent(CARD_ACTIVATED_TOPIC, {
       userId,
       customerId,
     });
