@@ -1,2 +1,8 @@
 export * from './lib/ltv/ltv.service';
-export * from './lib/lib.module';
+export * from './lib/ltv/ltv.module';
+export * from './lib/ltv/ltv.controller';
+export * from './lib/ltv/ltv.dto';
+
+export * from './lib/margin/margin.dto';
+export * from './lib/margin/margin_calls.entity';
+export * from './lib/margin/utils/margin_action_handlers/margin_action_handlers.interfaces';
