@@ -1,18 +1,12 @@
 import { Module } from '@nestjs/common';
-import { CreditQueueController } from './credit.controller';
 import { CreditService } from './credit.service';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { LtvCollateral } from '../collateral.entity';
-import { LtvCredit } from '../credit.entity';
-import { UtilsModule } from '../utils/utils.module';
-import { CollateralTransaction } from '../collateral_transactions.entity';
+import { Credit } from './credit.entity';
+import { UtilsModule } from '../ltv/utils/utils.module';
 
 @Module({
-  controllers: [CreditQueueController],
-  imports: [
-    TypeOrmModule.forFeature([LtvCollateral, LtvCredit, CollateralTransaction]),
-    UtilsModule,
-  ],
+  controllers: [],
+  imports: [TypeOrmModule.forFeature([Credit]), UtilsModule],
   providers: [CreditService],
   exports: [CreditService],
 })
