@@ -1,32 +1,40 @@
 import { Injectable } from '@nestjs/common';
 import { LtvDto, LtvStatus } from './ltv.dto';
 import { QueueService } from '@archie/api/utils/queue';
-import { LtvUtilService } from '../utils/ltv.service';
-import { CreditAssetUtilService } from '../utils/credit_asset.service';
-import { CollateralValue, CreditAssets } from '../utils/utils.interfaces';
-import { CollateralValueUtilService } from '../utils/collateral_value.service';
+import { LtvUtilService } from './utils/ltv.service';
+import { LedgerService } from '../ledger/ledger.service';
+import {
+  LedgerAccountUpdatedPayload,
+  LedgerActionType,
+} from '@archie/api/ledger-api/data-transfer-objects';
+import {
+  CreditBalanceUpdatedPayload,
+  PaymentType,
+} from '@archie/api/peach-api/data-transfer-objects';
+import { CreditLineCreatedPayload } from '@archie/api/credit-line-api/data-transfer-objects';
+import { CreditService } from '../credit/credit.service';
+import { Lock } from '@archie/api/utils/redis';
+import { MarginService } from '../margin/margin.service';
+import { LedgerAccount } from '../ledger/ledger_account.entity';
+import { LtvMeta } from '../margin/margin.interfaces';
+import { LTV_UPDATED_TOPIC } from '@archie/api/ltv-api/constants';
 
 @Injectable()
 export class LtvService {
   constructor(
     private queueService: QueueService,
-    private creditAssetsUtilService: CreditAssetUtilService,
     private ltvUtilService: LtvUtilService,
-    private collateralValueUtilService: CollateralValueUtilService,
+    private ledgerService: LedgerService,
+    private creditService: CreditService,
+    private marginService: MarginService,
   ) {}
 
   async getCurrentLtv(userId: string): Promise<LtvDto> {
-    const creditAssets: CreditAssets =
-      await this.creditAssetsUtilService.getCreditInfo(userId);
+    const ltvMeta = await this.getNormalizedLtvMeta(userId);
 
-    const collateralValue: CollateralValue =
-      this.collateralValueUtilService.getCollateralValue(
-        creditAssets.collateral,
-        creditAssets.assetPrices,
-      );
     const ltv: number = this.ltvUtilService.calculateLtv(
-      creditAssets.credit.utilizationAmount,
-      collateralValue.collateralBalance,
+      ltvMeta.creditUtilization,
+      ltvMeta.ledgerValue,
     );
     const ltvStatus: LtvStatus = this.ltvUtilService.getLtvStatus(ltv);
 
@@ -34,5 +42,85 @@ export class LtvService {
       ltv,
       status: ltvStatus,
     };
+  }
+
+  @Lock((ledger: LedgerAccountUpdatedPayload) => ledger.userId)
+  async handleLedgerAccountUpdatedEvent({
+    userId,
+    ledgerAccounts,
+    action,
+  }: LedgerAccountUpdatedPayload): Promise<void> {
+    await this.ledgerService.updateLedgerAccounts(userId, ledgerAccounts);
+
+    if (action.type === LedgerActionType.LIQUIDATION) {
+      await this.marginService.acknowledgeLiquidationCollateralBalanceUpdate(
+        action.liquidation.id,
+      );
+    }
+    const ltvMeta = await this.getNormalizedLtvMeta(userId);
+
+    const ltv: number = this.ltvUtilService.calculateLtv(
+      ltvMeta.creditUtilization,
+      ltvMeta.ledgerValue,
+    );
+
+    this.queueService.publishEvent(LTV_UPDATED_TOPIC, {
+      userId,
+      ltv,
+    });
+
+    await this.marginService.executeMarginCallCheck(userId, ltv, ltvMeta);
+  }
+
+  @Lock((credit: CreditBalanceUpdatedPayload) => credit.userId)
+  public async handleCreditBalanceUpdatedEvent(
+    credit: CreditBalanceUpdatedPayload,
+  ): Promise<void> {
+    await this.creditService.updateCreditBalance(credit);
+
+    if (credit.paymentDetails?.type === PaymentType.liquidation) {
+      await this.marginService.acknowledgeLiquidationCreditBalanceUpdate(
+        credit.paymentDetails.id,
+      );
+    }
+    const ltvMeta = await this.getNormalizedLtvMeta(credit.userId);
+    const ltv: number = this.ltvUtilService.calculateLtv(
+      ltvMeta.creditUtilization,
+      ltvMeta.ledgerValue,
+    );
+
+    this.queueService.publishEvent(LTV_UPDATED_TOPIC, {
+      userId: credit.userId,
+      ltv,
+    });
+
+    await this.marginService.executeMarginCallCheck(
+      credit.userId,
+      ltv,
+      ltvMeta,
+    );
+  }
+
+  private async getNormalizedLtvMeta(userId: string): Promise<LtvMeta> {
+    const updatedLedgerAccounts: LedgerAccount[] =
+      await this.ledgerService.getLedgerAccounts(userId);
+    const ledgerValue: number = this.ledgerService.getLedgerValue(
+      updatedLedgerAccounts,
+    );
+    const creditUtilization: number = await this.creditService.getCreditBalance(
+      userId,
+    );
+
+    return this.marginService.reducePendingLiquidationAmount(
+      userId,
+      creditUtilization,
+      ledgerValue,
+    );
+  }
+
+  public async handleCreditLineCreatedEvent({
+    userId,
+  }: CreditLineCreatedPayload): Promise<void> {
+    await this.creditService.createCreditBalance(userId);
   }
 }
