@@ -1,3 +1,5 @@
+import { Event } from '@archie/api/utils/queue';
+
 export const SERVICE_NAME = 'paypal-api';
 export const SERVICE_QUEUE_NAME = `${SERVICE_NAME}-queue`;
 
@@ -14,8 +16,10 @@ export enum ConfigVariables {
   PAYPAL_RETURN_URL = 'PAYPAL_RETURN_URL',
   AUTH0_DOMAIN = 'AUTH0_DOMAIN',
   AUTH0_AUDIENCE = 'AUTH0_AUDIENCE',
+  RUN_MIGRATIONS = 'RUN_MIGRATIONS',
 }
 
-export const PAYPAL_PAYMENT_RECEIVED_TOPIC = 'paypal.payment.received';
+export const PAYPAL_PAYMENT_RECEIVED_TOPIC =
+  new Event('paypal.payment.received', 1);
 
 export const PAYPAL_PAYMENT_CURRENCY = 'USD';
