@@ -1,23 +1,23 @@
 import { Controller } from '@nestjs/common';
-import { Subscribe } from '@archie/api/utils/queue';
+import { Subscribe } from '@archie/api/utils/queue/decorators/subscribe';
 import { SERVICE_QUEUE_NAME } from '@archie/api/peach-api/constants';
 import {
-  CREDIT_LIMIT_UPDATED_TOPIC,
+  CREDIT_LINE_UPDATED_TOPIC,
   CREDIT_LINE_CREATED_TOPIC,
-} from '@archie/api/credit-limit-api/constants';
+} from '@archie/api/credit-line-api/constants';
 import { PeachBorrowerService } from './loan.service';
 import {
   EMAIL_VERIFIED_TOPIC,
   KYC_SUBMITTED_TOPIC,
 } from '@archie/api/user-api/constants';
 import {
-  CreditLimitUpdatedPayload,
-  CreditLineCreatedPayload,
-} from '@archie/api/credit-limit-api/data-transfer-objects';
-import {
   EmailVerifiedPayload,
   KycSubmittedPayload,
 } from '@archie/api/user-api/data-transfer-objects';
+import {
+  CreditLineUpdatedPayload,
+  CreditLineCreatedPayload,
+} from '@archie/api/credit-line-api/data-transfer-objects';
 
 @Controller()
 export class PeachBorrowerQueueController {
@@ -28,7 +28,6 @@ export class PeachBorrowerQueueController {
   @Subscribe(
     KYC_SUBMITTED_TOPIC,
     PeachBorrowerQueueController.CONTROLLER_QUEUE_NAME,
-    { logBody: false },
   )
   async kycSubmittedHandler(payload: KycSubmittedPayload): Promise<void> {
     await this.peachService.handleKycSubmittedEvent(payload);
@@ -37,7 +36,6 @@ export class PeachBorrowerQueueController {
   @Subscribe(
     EMAIL_VERIFIED_TOPIC,
     PeachBorrowerQueueController.CONTROLLER_QUEUE_NAME,
-    { logBody: false },
   )
   async emailVerifiedHandler(payload: EmailVerifiedPayload): Promise<void> {
     await this.peachService.handleEmailVerifiedEvent(payload);
@@ -54,12 +52,12 @@ export class PeachBorrowerQueueController {
   }
 
   @Subscribe(
-    CREDIT_LIMIT_UPDATED_TOPIC,
+    CREDIT_LINE_UPDATED_TOPIC,
     PeachBorrowerQueueController.CONTROLLER_QUEUE_NAME,
   )
-  async creditLimitUpdatedHandler(
-    payload: CreditLimitUpdatedPayload,
+  async creditLineUpdatedHandler(
+    payload: CreditLineUpdatedPayload,
   ): Promise<void> {
-    await this.peachService.handleCreditLimitUpdatedEvent(payload);
+    await this.peachService.handleCreditLineUpdatedEvent(payload);
   }
 }
