@@ -8,10 +8,7 @@ import { PurchasesResponseFactory } from './utils/purchases_response.factory';
 import { GetPurchasesQueryDto, PurchasesResponseDto } from './purchases.dto';
 import { Injectable } from '@nestjs/common';
 import { TransactionUpdatedPayload } from '@archie/api/credit-api/data-transfer-objects';
-import {
-  CreditBalanceUpdatedPayload,
-  PaymentType,
-} from '@archie/api/peach-api/data-transfer-objects';
+import { PaymentType } from '@archie/api/peach-api/data-transfer-objects';
 import { CREDIT_BALANCE_UPDATED_TOPIC } from '@archie/api/peach-api/constants';
 import { QueueService } from '@archie/api/utils/queue';
 
@@ -66,19 +63,16 @@ export class PurchasesService {
         borrower.personId,
         borrower.creditLineId,
       );
-      this.queueService.publish<CreditBalanceUpdatedPayload>(
-        CREDIT_BALANCE_UPDATED_TOPIC,
-        {
-          ...credit,
-          userId: transaction.userId,
-          paymentDetails: {
-            type: PaymentType.purchase,
-            amount: transaction.us_dollar_amount,
-            asset: 'USD',
-            id: String(transaction.id),
-          },
+      this.queueService.publishEvent(CREDIT_BALANCE_UPDATED_TOPIC, {
+        ...credit,
+        userId: transaction.userId,
+        paymentDetails: {
+          type: PaymentType.purchase,
+          amount: transaction.us_dollar_amount,
+          asset: 'USD',
+          id: String(transaction.id),
         },
-      );
+      });
     } else {
       await this.peachApiService.updatePurchase(
         borrower.personId,
