@@ -1,3 +1,6 @@
+import { Event } from '@archie/api/utils/queue';
+import { CreditBalanceUpdatedPayload } from '@archie/api/peach-api/data-transfer-objects';
+
 export const SERVICE_NAME = 'peach-api';
 export const SERVICE_QUEUE_NAME = `${SERVICE_NAME}-queue`;
 
@@ -18,10 +21,10 @@ export enum ConfigVariables {
   ENCRYPTION_KEY = 'ENCRYPTION_KEY',
   API_BASE_URL = 'API_BASE_URL',
   REDIS_URL = 'REDIS_URL',
+  RUN_MIGRATIONS = 'RUN_MIGRATIONS',
 }
 
-export const CREDIT_BALANCE_UPDATED_TOPIC = 'peach.credit_balance.updated';
-export const AVAILABLE_CREDIT_LIMIT_BALANCE_UPDATED_TOPIC =
-  'peach.available_credit_limit_balance.updated';
+export const CREDIT_BALANCE_UPDATED_TOPIC =
+  new Event<CreditBalanceUpdatedPayload>('peach.credit_balance.updated', 1);
 
 export const GET_LOAN_BALANCES_RPC = 'get.loan.balances.rpc';
