@@ -1,3 +1,9 @@
+import { Event } from '@archie/api/utils/queue';
+import {
+  AppliedToWaitlistPayload,
+  JoinedToWaitlistPayload,
+} from '@archie/api/referral-system-api/data-transfer-objects';
+
 export const SERVICE_NAME = 'referral-system-api';
 export const SERVICE_QUEUE_NAME = `${SERVICE_NAME}-queue`;
 
@@ -10,10 +16,20 @@ export enum ConfigVariables {
   ARCHIE_MARKETING_WEBSITE_URL = 'ARCHIE_MARKETING_WEBSITE_URL',
   QUEUE_URL = 'QUEUE_URL',
   ENCRYPTION_KEY = 'ENCRYPTION_KEY',
+  RUN_MIGRATIONS = 'RUN_MIGRATIONS',
 }
 
-export const APPLIED_TO_WAITLIST_TOPIC = 'waitlist.joined';
+export const APPLIED_TO_WAITLIST_TOPIC = new Event<AppliedToWaitlistPayload>(
+  'waitlist.applied',
+  1,
+);
 
-export const JOINED_WAITLIST_TOPIC = 'waitlist.joined';
+export const JOINED_WAITLIST_TOPIC = new Event<JoinedToWaitlistPayload>(
+  'waitlist.joined',
+  1,
+);
 
-export const SALES_CONNECT_TOPIC = 'sales.connect'
+export const SALES_CONNECT_TOPIC = new Event(
+  'sales.connect',
+  1,
+);
