@@ -1,7 +1,7 @@
 import { JwtPayload, sign, verify } from 'jsonwebtoken';
 import { user } from './stubs/user.stubs';
 
-export const generateUserAccessToken = (customUser = user) => {
+export const generateUserAccessToken = (customUser = user): string => {
   return sign({}, 'ACTIONS_SIGNING_SECRET', {
     expiresIn: 300,
     issuer: 'AUTH0_DOMAIN',
