@@ -3,6 +3,6 @@ import { QueueUtilService } from '@archie/api/utils/queue';
 export const GLOBAL_EXCHANGE_NAME = QueueUtilService.GLOBAL_EXCHANGE.name;
 
 export const queueStub = {
-  publish: jest.fn(),
+  publishEvent: jest.fn(),
   request: jest.fn(),
 };
