@@ -14,7 +14,8 @@ export class Auth0Service {
       clientSecret: this.configService.get(
         ConfigVariables.AUTH0_M2M_CLIENT_SECRET,
       ),
-      scope: 'read:users update:users create:guardian_enrollment_tickets',
+      scope:
+        'read:users update:users create:guardian_enrollment_tickets delete:guardian_enrollments',
     });
   }
 
