@@ -1,2 +1,2 @@
-export * from './lib/user.module';
-export * from './lib/user.interfaces';
+export * from './lib/email/email.module';
+export * from './lib/mfa/mfa.module';
