@@ -1,12 +1,16 @@
+import { InjectionToken, OptionalFactoryDependency } from '@nestjs/common';
+import { Type } from '@nestjs/common/interfaces/type.interface';
+import { DynamicModule } from '@nestjs/common/interfaces/modules/dynamic-module.interface';
+import { ForwardReference } from '@nestjs/common/interfaces/modules/forward-reference.interface';
+
 export interface CryptoConfig {
   encryptionKey?: string;
 }
 
 export interface CryptoOptions {
-  // eslint-ignore-next-line
-  imports: any[];
-  // eslint-ignore-next-line
-  inject: any[];
-  // eslint-ignore-next-line
-  useFactory: (...args: any[]) => CryptoConfig;
+  imports: Array<
+    Type | DynamicModule | Promise<DynamicModule> | ForwardReference
+  >;
+  inject: (InjectionToken | OptionalFactoryDependency)[];
+  useFactory: (...args: unknown[]) => CryptoConfig;
 }
