@@ -2,13 +2,20 @@ import { INestApplication, ValidationPipe } from '@nestjs/common';
 import { HttpAdapterHost, NestFactory } from '@nestjs/core';
 import { WinstonModule } from 'nest-winston';
 import * as winston from 'winston';
-import { Openapi } from '@archie/api/utils/openapi';
 import { AllExceptionsFilter } from '@archie/api/utils/tracing';
 import { BigNumber } from 'bignumber.js';
+import { WsAdapter } from '@archie/api/websocket-event-api/websocket-event';
+
+export interface StartServiceOptions {
+  enableWs?: boolean;
+}
 
 export async function startService(
   _name: string,
   module: unknown,
+  options: StartServiceOptions = {
+    enableWs: false,
+  },
 ): Promise<INestApplication> {
   const app = await NestFactory.create(module, {
     logger: WinstonModule.createLogger({
@@ -22,20 +29,22 @@ export async function startService(
         }),
       ],
     }),
+    rawBody: true,
   });
   BigNumber.config({
     DECIMAL_PLACES: 18,
     ROUNDING_MODE: BigNumber.ROUND_DOWN,
     EXPONENTIAL_AT: 19,
   });
-
-  await Openapi.generate(app);
-
   const { httpAdapter } = app.get(HttpAdapterHost);
 
   app.useGlobalFilters(new AllExceptionsFilter(httpAdapter));
   app.useGlobalPipes(new ValidationPipe({ whitelist: true }));
   app.enableCors();
+
+  if (options.enableWs === true) {
+    app.useWebSocketAdapter(new WsAdapter(app));
+  }
 
   await app.listen(process.env.PORT ?? 80);
 
