@@ -1,71 +1,32 @@
 import { Injectable } from '@nestjs/common';
-import { PeachApiService } from './api/peach_api.service';
-import { InjectRepository } from '@nestjs/typeorm';
-import { Repository } from 'typeorm';
-import { PeachEvent } from './peach_events.entity';
-import { EventsResponse, Payment } from './api/peach_api.interfaces';
 import { QueueService } from '@archie/api/utils/queue';
-import { WEBHOOK_PEACH_PAYMENT_CONFIRMED_TOPIC } from '@archie/api/webhook-api/constants';
-import { WebhookPaymentPayload } from '@archie/api/webhook-api/data-transfer-objects';
+import { WEBHOOK_PEACH_PAYMENT_UPDATED_TOPIC } from '@archie/api/webhook-api/constants';
+import { PeachWebhookPayload } from '@archie/api/webhook-api/data-transfer-objects';
 
 @Injectable()
 export class PeachWebhookService {
-  PAYMENT_CONFIRMED_EVENT_UUID = 'ab4012cc-8042-4911-90d4-85cb32b665e5';
+  constructor(private queueService: QueueService) {}
 
-  constructor(
-    @InjectRepository(PeachEvent)
-    private peachEventRepository: Repository<PeachEvent>,
-    private peachApi: PeachApiService,
-    private queueService: QueueService,
-  ) {}
+  public async handlePaymentEvent(payment: PeachWebhookPayload): Promise<void> {
+    const transaction = payment.transaction;
 
-  public async publishPaymentConfirmedEvent(body: WebhookPaymentPayload) {
-    this.queueService.publish<WebhookPaymentPayload>(
-      WEBHOOK_PEACH_PAYMENT_CONFIRMED_TOPIC,
-      body,
-    );
-  }
-
-  // This is just a temporary solution until Peach implements webhooks
-  // Currently we just poll for new events
-  public async handlePaymentConfirmedEvent(): Promise<void> {
-    const peachEvent: PeachEvent | null =
-      await this.peachEventRepository.findOneBy({
-        uuid: this.PAYMENT_CONFIRMED_EVENT_UUID,
-      });
-
-    const paymentEvents: EventsResponse<Payment> =
-      await this.peachApi.getPaymentConfirmedEvent(
-        peachEvent?.lastFetchedPaymentConfirmedEventId,
-      );
-
-    paymentEvents.data.forEach((event: Payment) => {
-      this.queueService.publish<WebhookPaymentPayload>(
-        WEBHOOK_PEACH_PAYMENT_CONFIRMED_TOPIC,
-        {
-          ...event,
-          amount: event.amount / 100, // Peach returns amount in cents eg. 1100 for 11.00$
+    this.queueService.publishEvent(WEBHOOK_PEACH_PAYMENT_UPDATED_TOPIC, {
+      userId: payment.borrower.externalId,
+      borrowerId: payment.borrowerId,
+      loanId: payment.loanId,
+      transaction: {
+        id: payment.transactionId,
+        actualAmount: transaction.actualAmount,
+        paymentDetails: {
+          fromInstrumentId: transaction.paymentDetails.fromInstrumentId,
         },
-      );
+        status: transaction.status,
+        drawId: transaction.drawId,
+        externalId: transaction.externalId,
+        metaData: transaction.metaData,
+        parentTransactionId: transaction.parentTransactionId,
+        currency: payment.currency,
+      },
     });
-
-    if (paymentEvents.data.length > 0) {
-      const lastEvent: Payment =
-        paymentEvents.data[paymentEvents.data.length - 1];
-
-      await this.peachEventRepository.upsert(
-        {
-          uuid: this.PAYMENT_CONFIRMED_EVENT_UUID,
-          lastFetchedPaymentConfirmedEventId: lastEvent.id,
-        },
-        {
-          conflictPaths: ['uuid'],
-        },
-      );
-
-      if (paymentEvents.nextUrl !== null) {
-        await this.handlePaymentConfirmedEvent();
-      }
-    }
   }
 }
